Merged PR 6185425: Bug 14181716: Listing packages for a repo raises 500 error
This fixes the immediate cause of the error (missing type arg).
The deeper problem is that we are intetionally subverting arg and type
checking for everything that implements the PulpApi interface, but
that would be a much larger change and it's not immediately obvious
to me what we should do instead.

Related work items: #14181716