Merged PR 6211573: TASK 14708576, 14491466, 14722525: Improve test coverage
This adds happy-path test coverage for all our commands (except 'task cancel', which I don't think we should bother with). We're not actually trying to test that Pulp is doing everything correctly, merely that our code is passing the expected things through, so this is probably good enough for a test suite for now.

Along the way I also fixed tasks 14491466 and 14722525 since these were blockers I encountered, and added another arg to the to the 'orphan cleanup' command that I needed. I'm fine with all this being squash merged if you all are.

Related work items: #14491466, #14673107, #14708576, #14722525