Merged PR 6285237: Make authentication default for all routes
This PR does nothing.

Except flip the paradigm so that instead of routes being unauthenticated by default (so we need to annotate each one), now they are authenticated by default (and we have to specify any path that we do not want to have to authenticate).

This protects us from accidentally forgetting to add authentication to some new route, which is a type of problem that can be hard to detect if you're just looking at the diff in a review.

Also handles 401s in the client more gracefully than spouting a json parse error.

Related work items: #14877632