Merged PR 6298942: If command raises, emit error body to stderr
When pmcserver returns a non-2xx status code, additional information about the reason appears in the JSON body of that response. The cli's use of httpx raises an httpx.HTTPStatusError in that case, and the catch block in main.py:run() emits the JSON body. Some common usage patterns for the CLI use --id-only and capture the output of the pmc command into a variable, so emitting the JSON error body to stderr effectively makes it invisible.