Merged PR 6410526: Create a Pre-Production Environmnet
This deployment script and yaml files creates an AKS cluster, and ACR instance, and a Postgresql server contained in a vnet for privacy.

The Pulp instance we're currently using is not sufficient long-term, we need to build it ourselves both for supply-chain security reasons and to integrate with Azure Blob Storage for the filesystem, but it's good enough for now.

The environment that this creates should be highly-available and secure.

Related work items: #14617110