Merged PR 6669612: Clean up a couple things that weren't using LIST_SEPARATOR
The `access` command was being created concurrently with the PR that defined `LIST_SEPARATOR`, and as a result didn't use it. Also we apparently missed the distro package update command. This PR just cleans up both places so that we consistently use semicolon-separated lists in the cli.

Related work items: #15313841