Merged PR 6664804: Migrate to a 3-container pulp dev installation
Right now our dev environment is using the "non-production-ready" "all-in-one" pulp image from `docker.io/pulp/pulp` for simplicity and ease of setup. However this needs to change now, because we are soon going to be building our own pulp image that we will need to test/use, and that's going to require the three-container-with-different-roles setup.

Migrate our docker-compose setup to configure and use a three-container installation of pulp based on the `quay.io/pulp/pulp` image.

This is a significant update to the `.env` file and the database configuration (pulp's database is now co-located in the db container, not a locally-running postgres instance), so you will need to completely wipe and re-create your environment to apply this change. There is now a helpful make command to do that: `make reset_rebuild`

Related work items: #15374541