Merged PR 6712285: Improve security by dropping AKS Secrets
This is a nice-to-have improvement that allows us to drop the step where we sync KeyVault secrets into AKS Secrets so they can be added into env variables. Instead we have scripts that can read the secrets from the filesystem directly into the env variables, allowing us to cut out the AKS middleman.

Also change the order of the KeyVault creation to lock it down more tightly *before* we add secrets to it.

Related work items: #15447622