Merged PR 6731677: Allow Pulp to sign repodata, update pmcserver / cli to allow you to specify signing service
At a "10,000 ft view" the way repodata signing is supposed to work in Pulp is that you have some local script that Pulp can call out to, you register it with `pulpcore-manager --add-signing-service`, and then you reference its `pulp_href` url when creating a repo (yum) or when publishing content (apt). Additionally yum/apt repos want different signing services that do slightly different things. If you have it all set up properly then Pulp will sign and publish signatures as part of the `publish` step. To make our lives easier I am storing the desired singing service in a pulp label for apt repos and referencing it every time we publish apt content in order to present a unified interface.

In our case, we have two fundamentally different types of signing that we need to be able to do. "Legacy" signing is done locally in the `signer` container using the MS Open Tech key. "ESRP" signing authenticates with an auth cert and then calls out to `az xsign` to get ESRP to sign the files with the secret key that only they have access to. There is also a "test ESRP" signing process that uses a different key that the PPE is expected to be using to test things. Most of the code to do this was already present in the `signer` container and `sign_cli` scripts, except that Matt had foolishly trusted the Pulp documentation and ripped out all the code dealing with clearsigning (instead of *only* detached signing), which turns out to be necessary to sign apt repodata. So I had to add that back in, and along the way I noticed a couple of logic sub-optimalities and I found the old code fairly confusing, so I pretty heavily re-factored `signer.py` to make more sense to me and hopefully be more robust. (For example, the code in vCurrent will retry *both* the clearsign and the detached-sign operations if the second one fails with a recoverable error. Since these are potentially both very slow operations, ideally we'd like to only retry the failed attempt.)

I've updated our dev setup to include the `signer` container, complete with a locally-generated key so that we can do "legacy" signing with that key. As a result there are a fairly large amount of environment updates, so you'll need to remake your env with `make reset_rebuild` to pull them in once this is merged. "ESRP" signing is not available in dev environments unless you happen to have an auth cert on you and want to muck around with setting the appropriate env variables.

I am unable to test ESRP signing in dev, so it's possible that I messed something up in `signer` during the refactoring, although I was trying to be careful. I also have not tried pushing these changes into the PPE yet, so the deployment configs might need some additional work.

I'll go section-by-section to describe the changes, but that'll have to be in the first comment because I'm hitting the size limit.

Related work items: #15477573