Merged PR 6784842: Make oid editable on Account
This PR adds a new `oid` field to Account and changes the semantics of `id` so that it's a randomly-generated uuid instead of the `oid`. The upshot of this is that we can change `oid` on an account if we need to (someone leaves the company, someone else takes over with a different principal...) without breaking foreign key references.

NOTE: This db migration will wipe all data in Account, RepoAccess, and OwnedPackage. We would have to re-import account / permission information in the PPE. If that's not acceptable let me know.

Related work items: #15548198