Merged PR 6796402: Add new api/command for cloning repo permissions
We want some way to easily be able to clone repo permissions from an existing repo to a new one that's owned by the same people. I decided to do this as a new command / api method so that we don't *only* **have** to do it when creating the repo.

Related work items: #15354092