Merged PR 6809776: Accidental package removal from apt repo
The point of this PR is to fix the edge case in [15480205](https://msazure.visualstudio.com/One/_workitems/edit/15480205) wherein you can accidentally break other apt distro/releases in this repo if you remove a package from any one distro/release, because it also removes the package from the whole repo.

Along the way I discovered that my code changes to remove a package from a single distro/release were not working because of an apparent bug in `pulp-deb`, filed as [15641546](https://msazure.visualstudio.com/One/_workitems/edit/15641546) and not yet fixed. Instead we decided to move ahead with the code as-is, forcing users to remove the package from the entire repo explicitly (and they can add it back in to specific releases later if they want.

To describe the code changes, I:
1. Refactored the logic for managing apt release/component mappings out of `services.pulp.api` and into its own class. IMO the api portion should be fairly thin and deal with translating api/arg names, and this had a fair bit of logic in it.
1. Documented via docstring what `pulp_deb` expects and how it should work.
1. Added unit tests.
1. Updated the logic to both fix this bug (if the `pulp_deb` bug gets fixed) and also allow for an explicit "remove this package from all releases in this repo" command, which was not previously possible.

Related work items: #15480205