Merged PR 6804195: Add commands to create storage accounts (pulp artifacts, access logs)
Pulp needs a storage account to hold the artifacts it's managing, and Azure Analytics needs a storage account to which it can write access logs for read/write/delete ops on the artifact storage account. In this PR adds commands to:

- create the storage accounts
- create the "pulp" container in the artifact storage account
- enable Azure Analytics on the artifact storage account to write logs to the other storage
- add the active blob storage key to keyvault as a secret

This PR does not add the artifact storage account to keyvault to enable rotation, as keyvault doesn't support extracting the active key for a storage account it managed. Instead, a subsequent PR will add a script to rotate the keys and update keyvault.

Related work items: #15628868