Merged PR 6814629: Initialize db the same way in AKS and docker.
There was a chicken-and-egg problem with the environment deployment scripts where we needed to shell out to a container to do the initial db setup, but the containers would die if the databases weren't available.

In our dev environments this was being done automagically because we were injecting a sql script into the postgres container that it would execute on creation.

This PR unifies the process in both environments so that the pmcserver migrate "job" does the initial db creation if not created already, and ensures that that "job" runs first.

Also I added a heathcheck on the docker postgres container and made things depend on `SERVICE_HEALTHY` instead of `SERVICE_STARTED`, because there was a problem where things would try to connect to the database too quickly and die because it wasn't up yet.