Merged PR 6834315: Fix pulp filters, allow single release changes
It turns out that there was no bug preventing `pulp_deb` from removing a package from a single release, it was my `pulp_deb` list filters that were wrong. This PR updates them so they work as expected, with a longer discussion in the comments of `pulp_deb` itself for posterity.

Related work items: #15641546