Merged PR 6898524: Delete local AKS volume and fix a couple small prod config bugs
The local AKS storage volume is no longer used now that everything is going to the storage account. Also the values of ConfigMaps should always be strings or it will whine at you, and my previous PR had a mistake where I overrode the name of `aks` instead of `kv`.