Merged PR 6919211: Terminate ssl connections in AKS properly
This adds an nginx container that terminates the incoming ssl connections and forwards traffic on to the appropriate :80 services. There are a couple of new internal-only services that point at the pmc api and pulp-content that nginx can forward to, and the externally-addressable services are now pointing to nginx.

In the process of testing this I deleted and re-created the external services in ppe, so the ip addresses there have changed. Furthermore it seems to take a while for the new ones to propagate, so I'm not able to hit the new addresses from my workstation yet. But at least one of them is available from inside the AKS cluster already, and if we fake out a DNS entry to get `curl` to send the appropriate HOST header to our new external endpoints then things seem to work:

```
curl --resolve csd.packages.ppe.trafficmanager.net:443:20.241.233.121 https://csd.packages.ppe.trafficmanager.net/pulp/content/

        <!DOCTYPE html>
        <html>
            <body>
                <ul>

                    <li><a href="repos/">repos/</a></li>

                    <li><a href="yumrepos/">yumrepos/</a></li>

                </ul>
            </body>
        </html>
```

```
curl --resolve csd-apt-cat-ppe.westus2.cloudapp.azure.com:443:20.241.233.121 https://csd-apt-cat-ppe.westus2.cloudapp.azure.com/api/

{"server":{"version":"0.0.1"},"versions":{"v4":"v4/"}}
```

Is the fact that nginx keys off of the HOST header and not the incoming ip address a problem Matt? Note in the examples above I'm hitting the same ip address both times (with different HOST headers), and both the api and pulp-content are accessible.