Merged PR 6941809: Fix api endpoint port
I merged the ssl-endpoint PR too eagerly.

1. Fix for public api endpoint port.
1. Switch the order of the nginx server definitions, so that if no HOST header is provided it will hit pulp-content and not the API.

The behavior here is a little weird, even though there are two IPs / LoadBalancers, supposedly one for the api and one for pulp-content, since it's going through one reverse proxy if the HOST header is not provided then nginx doesn't know which service you're trying to hit, so it'll just forward you to the first one defined. We don't want that first-defined service to be the api so that customers won't accidently hit it / discover it if they try to browse by ip address (with https).

The reverse is still true, Publishers trying to hit the API will hit pulp-content and get errors if they are using the ip address directly instead of the dns name. But oh well, they'll just have to use the dns name.