Merged PR 7013154: Fix empty apt repos
We apparently should not be manually creating the releases before we sync apt repos from vCurrent, it leads to a "duplicate Release" error and the sync fails.

We will have to wipe all apt remotes / distros / repos and re-create them after this is merged, so I'm also including a helper script to do that first part.

Related work items: #16002718