Merged PR 7030736: Fix 401 responses while polling tasks
Prior to this change, we were acquiring the token once and setting it in
the client. This change calls acquire_token() each time a request is
made.

It's worth noting that acquire_token() is calling the python MSAL
package's acquire_token_silent() function which first attempts to pull
the token from cache if it hasn't expired.

fixes #16040263

Related work items: #16040263