Merged PR 7029709: Increase function timeouts to prevent errors
We're encountering MessageLockLost exceptions which seem to be because
we're not locking the service bus queue for long enough. The default is
5 but we need to lock it for the duration of the function execution.

Also, increase the function duration from the default of 30 to 60
minutes to be safe. I don't expect syncs to take that long but we need
to account for possible task queue times.