Merged PR 7079387: Use an older style for older pygments versions
Use an older style for older pygments

Pygments is an optional dependency in pmc-cli that provides color
output. As an optional dependency, a version is not required. The
version on Ubuntu 20.04 is 2.3.1 which doesn't have the solarized
styles so use an older style when if solarized is not available.

fixes #16141884

Related work items: #16141884