Merged PR 7101622: Publishing / signing repodata is too slow for azure function to wait for
We currently have a (growing) backlog of tasks in the Azure Function queue, partially because the function waits for the publish/sign task which is currently [extremely slow](https://msazure.visualstudio.com/One/_workitems/edit/16101470). This PR tells it not to wait so it can keep up better with the changes in vCurrent.

Related work items: #16186329