Merged PR 7118952: process_actions can filter packages by repo
Allow the functions to specify the repo the package is in to avoid duplicates. Also add the options to the cli because why not.

The root problem is that we are explicitly allowing duplicate NEVRA (rpm) and NVA (deb) with different `relative_path`s so that we can sync content from vCurrent, where it is only enforced that they're unique *per repo*. So if we have the function specify the repo too then we should be able to avoid the ambiguity problem, and only one package should be returned.

Related work items: #16219095