Merged PR 7142787: Increase cpu / number of nginx containers
For nginx increase the max cpu to 1 whole cpu and the min number of nodes to 4 to handle the spiky 15-minute load cycle without 502s.

I'm also transitioning to defining the HorizontalPodAutoscalers declaratively in the yml files instead of having them created once on the kubectl command line. This way we're actually able to update them by changing things in the yml files and applying them. The part is a no-op change from what is in prod currently,

Related work items: #16261568