Merged PR 7161133: Disable prefetch and set maxMessageBatchSize
Prefetching is disabled by default. We enabled it as I thought it
controlled batch sizes but now I see it's an independent feature and is
likely the cause of the message lock exceptions we're seeing.

From https://learn.microsoft.com/en-us/azure/service-bus-messaging/service-bus-performance-improvements?tabs=net-standard-sdk-2#prefetching:

The prefetched copy of the message remains in the cache. The receiver
that consumes the expired cached copy will receive an exception when it
tries to complete that message. By default, the message lock expires
after 60 seconds. This value can be extended to 5 minutes. To prevent
the consumption of expired messages, set the cache size smaller than the
number of messages that a client can consume within the lock timeout
interval.

As our function can take longer than 5 min, it's likely we'll encounter message
lock exceptions if we attempt to use prefetching.

fixes #16298472

Related work items: #16298472