Merged PR 7174536: Accept default settings from the 'default' profile if not overridden
Accept default settings from the "default" profile if not overridden in other profiles.

There is a slight change of semantics here where previously other profiles would accept the "automatic" default of any value not set, which is not true any more if the "default profile" sets it. You would instead have to explicitly override the value in the default profile.