Merged PR 7202114: Extend graceful shutdown period for Pulp Tasks to up to 5 minutes.
Both Docker and AKS signal that they want to shut a container down by sending SIGTERM, and then waiting a bit. Let's set the config values such that Pulp tasks have up to 5 minutes to complete successfully and exit before we kill/cancel them.

There was also previously a problem in `signer` where it was not propagating the SIGTERM properly, which is why it always takes the max timeout to shut down currently. It should now propagate properly, which means we'll actually shutdown faster on average. There is still a timing window here where we could receive the SIGTERM at the very beginning of a `publish` task and we have not yet sent any signing requests to the `signer` container, so it will immediately shut down and won't be available when pulp tries to sign things and the task will error. I'm not sure how we can easily close that hole, but this should cover the 90% case.

For more discussion see the attached task.

Related work items: #16363508