Merged PR 7192414: Split nginx container, make min number of workers env configrable
The root problem here is that the api traffic was previously only forwarded based off of hostname which is not the most robust setup, and in PPE the desired hostname is not currently forwarding traffic to AKS. If we duplicate the nginx setup, with one pod to handle content traffic and another to handle api traffic, then we can keep track of where every request needs to go regardless of hostname.

I'm also making the min number of containers configurable, since there's no reason that PPE needs 10 workers like prod is currently configured, and adding a hidden `ssl_verify` option so you can turn ssl verification off if you're (for example) pointing at the PPE IP address directly which is not listed in the cert.

There would also be a 1-time migration necessary to clean up the orphaned generic autoscaler and deployment:
```
kubectl delete hpa nginx
kubectl delete deployment nginx
```