Merged PR 7411702: Enable postgres cert verification in pulp too
This update causes Pulp to tell the postgres cli to validate the server's cert against the rood Digicert cert that it's supposed to be signed with, preventing man-in-the-middle attacks. We do this already in pmcserver, but at the time we weren't building our own Pulp images. But we are now, so we can do it here too.

Related work items: #15349935