Merged PR 7623329: Makefile/setup fixes
Some fixes to the Makefile and setup docs. First, have the `.env` target
generate .test.key since it needs it. Also, use single colon so that we
don't recreate .env if it exists (the `rebuild_env` target exists for
this purpose). Lastly, update docs as `register_sigining_service` target
doesn't exist and we no longer have to manually call the generate
.test.key target.