Merged PR 7759746: Fix failing test, do consistency check first
Fix failing `tests/commands/test_repository.py::test_yum_update_packages_with_release` test. We had a check for this in place previously, but with recent code changes it now comes too late, and something tries to interpret the yum RepoId as an apt RepoId and it blows up. Move the check to the beginning of the api methods.