Merged PR 7786045: Provide a way to clear the deadletter queue
This is to enable the initial deployment of the updated migration functions. I have not found a satisfactory way to clear out the existing deadletter queue so that we don't file ICMs / permanently have these resolved messages in the system. So code up some custom way to clear it instead. I'll deploy the functions once with the env variable set, and then again when the queue is cleared with it not.

Related work items: #16225853, #16515579