Merged PR 7870323: Fix inconsistent return of format of id only results
The `--id-only` option currently returns either a single id or a json list of
ids. This is rather inconsistent and causes publishers to have to deal
with two possible formats.

The original purpose of the option was to avoid having users parse json.
Always returning a comma-separated list of ids would allow users to pass
the id(s) to other commands without having to parse json.

fixes #17749140

Related work items: #17749140