Merged PR 7850749: Adding the support to publish source packages to repos
Adding the support to publish source packages to repos.
The main change is that we assumed that any repo would only have one package type, Having the support for source packages, apt repos can now have deb and deb_src packages.
A patch to pulp_deb is created as part of that PR to add the support for filtering the source packages using release, the same behavior for deb packages.
The PR adds the support for adding or removing packages of homogeneous types to a repo with one command.

Related work items: #13082573