Merged PR 7889226: Add package name to verification failure messages
When uploading multiple packages, if one fails verification, it's
impossible to identify which package failed. Originally, I attempted to
add the filename to the message in the CLI but the code currently just
prints the error it gets from the server. Luckily the server has the
filename so it was possible to add it to the exception.

fixes #17798615

Related work items: #17798615