Merged PR 6359126: Fix intermittent 401 issues
Publishers are regularly getting 401 responses from the PMC API. I was
able to reproduce and at least in my case the problem was due to a
httpx.ReadTimeout. I think that in general it's likely that intermittent
connection problems are to blame.

Part of the challenge with fixing this is that the
fastapi-microsoft-identity library we're using catches all exceptions
and just returns a 401 response which masks all sorts of possible
problems such as network connection problems and treats them as 401s:

https://github.com/425show/fastapi_microsoft_identity/blob/e98f1ff4a86e436b2d6d874738ff655fe1a63e54/fastapi_microsoft_identity/auth_service.py#L112-L113

Ideally, we'd open a PR to fix this and add some better network
resilience like a higher timeout and logic to retry failed requests.

However, the library appears to be unmaintained--the author has left MS
and my PRs against the project to fix various issues have gone
unacknowledged. I've even tried to reach out to the team at MS via Teams
but they don't respond.

This patch takes the relevant code and moves it into our codebase. I've
also replaced python-jose which appears unmaintained (last release was
June 2021) and replaced it with pyjwt.

Related work items: #17504967