Merged PR 7969902: Unmangle the nginx-api log format
Our deploy scripts pass the config ymls through `envsubst` to set all the environment-specific variables we use. However, that is currently conflicting with the nginx log format command and we're uploading mangled configs. There doesn't appear to be a great general solution to this problem, however we can simply give envsubst something to do (in this case, replace `${no_envsubst}`, which is undefined, with nothing) so that it'll leave the other variables alone.