Merged PR 7970281: Update nginx resource requests, and log api response time
We're currently wasting a ton of cpu/memory on our nginx reverse-proxies. They just don't need very much. Bump it down (with no hard limits) to free up resources on the nodes for other things that might actually use them.

Also let's log on nginx-api how long it took to respond to requests, and improve the log format in general.