Merged PR 7984328: Switched CLI from httpx to requests
Using httpx, we were experiencing problems where the cli would hang
indefinitely after uploading larger files. It appeared that the server
was responding but httpx somehow wasn't picking that up and just
continued to wait on a server response. Switching to the requests
package seems to fix the problem.

fixes #15947571

Related work items: #15947571