Merged PR 8039325: Convert from Identifier <-> pulp_href automagically
This makes `Identifier` smart enough to always instantiate the most-specific applicable subclass of itself, and teaches it to know the corresponding `pulp_href`. In `pulp/api.py` we can teach it to automagically convert from `Identifier` to `pulp_href` instead of having to do it manually a million places, and we can delete a couple of utility functions.

There is a similar refactor that I'd like to do that should obsolete all the `endpoint` function definitions, but this is big enough for one PR.

Related work items: #20540888