Merged PR 8017510: Signing the rows of the Account table and verifying the signature
Signing the rows of the Account table and verifying the signature.
Adding an alembic migration to add the signature column and sign the existing entries.
Signing happens in Key Vault where the key is stored while verification happens locally for better performance while having Key Vault verification as backup.
The PR also includes the changes to support a seamless process for the dev environment by generating a local key and using it for the signature/verification in the dev environment.
I used the more secure Elliptic Curve key and added the instructions to create it to the resource creation script.
Added a python script to help signing the initial dev account both in the dev and production environments.

In this PR, I aim to collect some feedback before I go ahead and test my changes in PPE.

Related work items: #13082554