Merged PR 8147097: Mount dev_assets into api container
This is needed because we're generating a key inside the dev_assets
folder to sign db entries. We ought to persist the key as commands like
'make rebuild' will recreate the container but not the volumes (and
hence the database) which would cause the key and database data to not
match.