Merged PR 8166962: Auto-delete permissions when account is deleted and speed up integration tests by keeping state
The integration tests currently do a tremendous amount of work simply setting up / tearing down database state in pulp for _every test_. This update switches the scope of the fixtures to "session", so they'll only be created / torn down once per test run. This means the tests have to be _slightly_ more careful about testing what they actually want to test in the face of possibly pre-existing database state (or explicitly request a new repo or whatever if they need it), but it cuts the time it takes to run the tests in half.

I ran into the slight problem of cleanup failing because it was trying to delete the account before the permissions were deleted. I think the correct behavior in the application is to chain-delete all related permissions when an account is deleted, so I've made that change too.