Merged PR 8167981: Cache jwt client in worker memory for half an hour
If we cache the jwt client in the worker's memory then its own internal token cache can actually work, and we don't have to re-auth on every single request. This cuts the time to run the integration tests by another 30%, plus is probably the right thing to do regardless so we're not DOSing the AAD servers.

If a TTL of 30 minutes is not appropriate let me know a better one.

Related work items: #24062068