Merged PR 8169132: Run integration tests in three threads with pytest-xdist
This splits up our integration tests into three groups and runs each group in parallel with pytest-xdist.

I won't be offended if we decide not to merge this, because it does make the test unreliable with random failures being possible if the tests happen to step on each other. However it seems to work _pretty_ well, I've done 10 test runs since I got the changes worked out and they've all been green, and it cuts the execution time almost in half again from about 3:40 currently to about 2:00.