Merged PR 8191889: Tell nginx to send tcp keepalive packets for 10 minutes
This should solve a problem where long client requests (eg Mariner large package uploads) would die at about the 5-minute mark with 104/504 errors, or eventually die at the 10-minute mark with the configured read timeout. Sending TCP KeepAlive packets keeps the connection fresh so that "something" in the middle doesn't kill it while we wait for the package to process.

The `so_keepalive` option on the listen directive instructs it to send keepalive packets to the client (after a minute, once every minute thereafter, max 10 times), and the `proxy_` directives control it's behavior for the upstream connections.

The `types_hash_max_size` directive is unrelated but clears a warning on startup about inefficient types parsing.

Related work items: #24104278