Merged PR 8193878: Make signature checking and auth verification async
This should solve the 502 issues that Publishers sometimes see under heavy load.

The root cause of the 502 is (usually?) that the pmc application is too busy to respond to a new request before nginx-api reaches its 60s connection timeout. The only thing I can see that could be that slow and blocking is the signature verification code, so if we make that properly async then there should never be a 60s period where the server is too busy to receive a new connection.

We can also use the newly-required `aioshutil` functionality to make the auth token verification async.

Related work items: #24095858, #24107040