Merged PR 8216081: Fix how we output errors
Previously, we were dumping a Python dict which isn't valid json (as it
uses single quotes instead of double) when tty = False.

Also, I realized that the json output was being sent to stderr instead
of stdout. I'm printing a human-readable error to stderr. We could do
more to improve the error message by using information from the detail
field but I'll save this for a future PR.

Related work items: #24164193, #24164413