Merged PR 7551598: Align with upstream vision of release signing service overrides
Previously we had implemented a pulp_deb patch locally (0008) where we tied the SigningService directly to an Apt Repository or Release via foreign keys to set the "default" for the repo/release. [We submitted it upstream](https://github.com/pulp/pulp_deb/pull/689), and they requested changes where instead of tying Release<->SigningService directly we instead do it via indirect reference to avoid a problem that they've had for forever where if you make any local changes to a Release then syncing from another repo creates a duplicate release. So that's what the pulp patch here is, it's kind of the second half of the 0008 patch that brings us in sync with what pulp_deb has merged upstream.

This PR also has server/cli support for setting that field so we don't have to do it via the pulp api directly, like we did last time when 0008 was first released.

Because of the database changes in pulp when we deploy this we'll have to reset the overrides on azurecore and azurecore-test:
```
pmc repo update azurecore-apt --add-release-signing-services 'bionic=legacy;xenial=legacy'
pmc repo update azurecore-test-apt --add-release-signing-services 'bionic=legacy;xenial=legacy;trusty=legacy'
```

Related work items: #16950389